% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalTree.R
\docType{data}
\name{compounds}
\alias{compounds}
\title{Compounds with mass ranges}
\format{A data frame with 23 rows and 3 variables}
\usage{
data(compounds)
}
\description{
A dataset containing the the mass ranges of chemical compounds with a 10 ppm tolerance.
The variables are as follows.
}
\details{
\itemize{
  \item name. name of the compound
  \item low. low bound of the mass
  \item high. high bound of the mass
}
}
\keyword{datasets}
