% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interval.R
\docType{class}
\name{Interval-class}
\alias{Interval-class}
\alias{Interval}
\title{Interval}
\description{
An S4 class to represent a named numeric interval.
}
\section{Slots}{

\describe{
\item{\code{start}}{the lower end of the interval}

\item{\code{end}}{the higher end of the interval}

\item{\code{key}}{the name of the interval}
}}

