% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalTree.R
\docType{class}
\name{IntervalTree-class}
\alias{IntervalTree-class}
\alias{IntervalTree}
\title{IntervalTree}
\description{
A S4 class to represent interval tree.
}
\section{Slots}{

\describe{
\item{\code{data}}{a dataframe providing the intervals to be stored in the interval tree. The columns are key, start and end of intervals}

\item{\code{root}}{the root list of the interval tree built upon the data}
}}

